+++ src/entity.rs
@@ -1,6 +1,8 @@
+use crate::Error;
 use chrono::NaiveDateTime;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
+use std::convert::TryFrom;
 
 /// Orthanc entity kinds (types).
 ///
@@ -12,6 +14,21 @@ pub enum EntityKind {
     Study,
     Series,
     Instance,
+}
+
+impl TryFrom<bytes::Bytes> for EntityKind {
+    type Error = Error;
+
+    fn try_from(value: bytes::Bytes) -> Result<EntityKind, Error> {
+        let s = String::from_utf8(value.to_vec())?;
+        match s.as_str() {
+            "Patient" => Ok(EntityKind::Patient),
+            "Study" => Ok(EntityKind::Study),
+            "Series" => Ok(EntityKind::Series),
+            "Instance" => Ok(EntityKind::Instance),
+            _ => Err(Error::new(&format!("Unknown entity kind: {}", s), None)),
+        }
+    }
 }
 
 /// A trait, that implements common methods for all entity kinds
@@ -329,6 +346,30 @@ mod tests {
     use super::*;
     use chrono::NaiveDate;
     use maplit::hashmap;
+
+    #[test]
+    fn test_entity_kind_try_from_bytes() {
+        assert_eq!(
+            EntityKind::try_from(bytes::Bytes::from_static(b"Patient")).unwrap(),
+            EntityKind::Patient
+        );
+        assert_eq!(
+            EntityKind::try_from(bytes::Bytes::from_static(b"Study")).unwrap(),
+            EntityKind::Study
+        );
+        assert_eq!(
+            EntityKind::try_from(bytes::Bytes::from_static(b"Series")).unwrap(),
+            EntityKind::Series
+        );
+        assert_eq!(
+            EntityKind::try_from(bytes::Bytes::from_static(b"Instance")).unwrap(),
+            EntityKind::Instance
+        );
+        assert_eq!(
+            EntityKind::try_from(bytes::Bytes::from_static(b"Foobar")).unwrap_err(),
+            Error::new("Unknown entity kind: Foobar", None)
+        );
+    }
 
     #[test]
     fn test_entity_trait_patient() {