+++ src/lib.rs
@@ -9,7 +9,7 @@
 //!
 //! ```ini
 //! [dependencies]
-//! orthanc = "0.7.0"
+//! orthanc = "0.8.0"
 //! ```
 //!
 //! ## Usage